requirements.txt
#python==3.9.7
matplotlib
python-dotenv
#dvc
pytorch==1.10.1
cudatoolkit==11.3.1
torchvision==0.11.2
pytorch-lightning==1.5.8
einops==0.4.0
pandas==1.4.0
#streamlit==1.3.1
wandb==0.12.9
hydra-core==1.1.1
